# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Harency ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/change_password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change Password",
                "description": "Change password for user",
                "operationId": "authChangePassword",
                "requestBody": {
                    "description": "Pass user current password and new password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "current_password",
                                        "example": "password"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "new_password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong credentials. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/get_profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Profile",
                "description": "Get user profile",
                "operationId": "authGetProfile",
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong invalid email or name. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/update_profile": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update Profile",
                "description": "Update profile for user",
                "operationId": "authUpdateProfile",
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "john"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "abc@gmail.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "format": "phone_number",
                                        "example": "1234567890"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "format": "country_code",
                                        "example": "91"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong invalid email or name. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/upload_image": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Upload Image",
                "description": "Upload Image",
                "operationId": "authUploadImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Image to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong format response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong Format. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/create_profile": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create Profile",
                "description": "Create profile for user",
                "operationId": "authCreateProfile",
                "requestBody": {
                    "description": "Pass user name and email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "john"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "abc@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong invalid email or name. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}